requirements.txt
asgiref==3.4.1
boto3==1.20.28
botocore==1.23.28
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
Django==4.0
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
factory-boy==3.2.1
Faker==11.3.0
freezegun==1.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
mysqlclient==2.0.3
packaging==21.3
Pillow==9.0.0
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.7
